requirements.txt
anyio==4.0.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1697725767277/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blosc2==2.0.0
boltons @ file:///croot/boltons_1677628692245/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorcet==3.0.1
comm==0.1.4
conda @ file:///croot/conda_1699303130622/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1700150901600/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
contourpy==1.1.0
cox==0.1.post3
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.11.0
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
dill==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-gpu==1.7.2
fastjsonschema==2.18.0
filelock==3.12.3
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.9.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
GPUtil==1.4.0
graphviz==0.20.1
grpcio==1.57.0
h5py==3.10.0
hiddenlayer==0.3
huggingface-hub==0.17.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.5
imgaug==0.4.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
json5==0.9.14
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
jupyterthemes==0.15.9
kiwisolver==1.4.5
lazy_loader==0.3
lesscpy==0.15.1
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
libmr==0.1.9
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.5
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
overrides==7.4.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.0
pandocfilters==1.5.0
param==2.0.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip-autoremove==0.10.0
platformdirs==3.10.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyct==0.5.0
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests @ file:///croot/requests_1690400202158/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robustness==1.2.1.post2
rpds-py==0.10.3
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.30.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tables==3.8.0
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.1.1
torchview==0.2.6
torchvision==0.15.2
torchviz==0.0.2
tornado==6.3.3
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
transformers==4.33.1
triton==2.0.0
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
wandb==0.15.10
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
xmltodict==0.13.0
zstandard @ file:///croot/zstandard_1677013143055/work
